refactor: Refactor reset password page loader and remove unnecessary imports
